Add a test for the 404 page when not logged in
