+++ package.json
@@ -49,6 +49,8 @@
     "cpy-cli": "^1.0.1",
     "dotenv": "^4.0.0",
     "dotenv-cli": "^1.4.0",
+    "enzyme": "^3.3.0",
+    "enzyme-adapter-react-16": "^1.1.1",
     "eslint": "^4.19.1",
     "eslint-plugin-graphql": "^1.5.0",
     "graphcool": "^0.9.2",
