+++ src/App.test.js
@@ -2,22 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { MockedProvider } from 'react-apollo/test-utils';
 import { MemoryRouter } from 'react-router-dom';
-import * as getEnvironmentVariableModule from './utils/getEnvironmentVariables';
-
-const localStorageMock = {
-  getItem: jest.fn().mockReturnValue(null),
-  setItem: jest.fn(),
-  removeItem: jest.fn(),
-  clear: jest.fn()
-};
-global.localStorage = localStorageMock;
-getEnvironmentVariableModule.default = jest.fn().mockReturnValue({
-  auth0ClientId: '',
-  auth0Domain: '',
-  graphcoolApi: ''
-});
-
-const App = require('./App').default;
+import App from './App';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
