+++ src/testShims.js
@@ -1,0 +1,11 @@
+const localStorageMock = {
+  getItem: jest.fn().mockReturnValue(null),
+  setItem: jest.fn(),
+  removeItem: jest.fn(),
+  clear: jest.fn()
+};
+global.localStorage = localStorageMock;
+
+global.requestAnimationFrame = callback => {
+  setTimeout(callback, 0);
+};
